requirements.txt
streamlit>=1.32.0
pytest>=8.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
